Implemented `When.Do(Action<Workflow>)`
this is an excellent step towards not needing branches explicitly. However, I think there's probably several bugs and things to consider as of yet. From here, I should put in place checks to ensure that branching into the inner code block requires a certain condition to be true. Perhaps I'll just let it be for now